Enable MDP/HWC composition
I measured around 2% extra benchmark performance and -10%  batery usage
on video playing.
Motorola seems that it made a mistake keeping this off, even oficial OTA
release has it on, while Qualcomm recomends it on.

However it should be checked for errors. Since I tested it on a CM based
rom.